refactor: create firestore handlers
