--- src/lib/leaveHandler.tsx
@@ -1,15 +1,0 @@
-import { functions } from '@/lib/firebase';
-import {
-  HttpsCallable,
-  httpsCallable,
-} from 'firebase/functions';
-
-export const leaveRoom = async (roomId: string) => {
-  console.log(roomId);
-  const leaveRoomCallback: HttpsCallable<{
-    roomId: string;
-  }> = httpsCallable(functions, 'leaveRoom');
-  await leaveRoomCallback({
-    roomId: roomId,
-  });
-};
